Adds `convert_migration_id` to ResourceMapService
LinkResolver relies on the migration_id_convertere to
implement a convert_migration_id method to handle
fallback cases where the link type is not specifically
handled.

It's likely the resource map wont actually have any...
but this at least adds future proofing and flexibility
if we need to add new link types in the future.